Okay, I may kind of see some problems. Weeeee

+++ NotThatGameEngine/NotThatGameEngine/Animation.h
@@ -47,11 +47,11 @@ public:
 private:
 
 	void UpdateGameObjectsTransform(const AnimationData* data);
-	void UpdateGameObjectsTransformRecursively(GameObject* gameObject, const AnimationData* data, int currentFrame);
+	void UpdateGameObjectsTransformRecursively(GameObject* gameObject, const AnimationData* data, float currentFrame);
 	void UpdateMesh(GameObject* gameObject);
-	float3 GetUpdatedChannelPosition(const Channels* channel, const int currentFrame) const;
-	Quat GetUpdatedChannelRotation(const Channels* channel, const int currentFrame) const;
-	float3 GetUpdatedChannelScale(const Channels* channel, const int currentFrame) const;
+	float3 GetUpdatedChannelPosition(const Channels* channel, const float currentFrame) const;
+	Quat GetUpdatedChannelRotation(const Channels* channel, const float currentFrame) const;
+	float3 GetUpdatedChannelScale(const Channels* channel, const float currentFrame) const;
 	void AnimateMesh(Mesh* mesh);
 
 public:
